feat: add reusable ConfirmDialog component
