+++ frontend/src/components/ConfirmDialog/index.ts
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const ConfirmDialog = React.lazy(() => import('./ConfirmDialog'));
